Removido Serviço Token Jwt
