+++ src/SME.SGP.Aplicacao/Queries/ObterPerfilDoToken/ObterPerfilDoTokenQuery.cs
@@ -1,0 +1,8 @@
+using System;
+using MediatR;
+
+namespace SME.SGP.Aplicacao;
+
+public class ObterPerfilDoTokenQuery : IRequest<Guid>
+{
+}
