+++ teste/SME.SGP.Aplicacao.Teste/Comandos/ComandosUsuarioTeste.cs
@@ -18,7 +18,6 @@ namespace SME.SGP.Aplicacao.Teste.Comandos
         private readonly Mock<IServicoAutenticacao> servicoAutenticacao;
         private readonly Mock<IServicoEol> servicoEOL;
         private readonly Mock<IServicoPerfil> servicoPerfil;
-        private readonly Mock<IServicoTokenJwt> servicoTokenJwt;
         private readonly Mock<IServicoUsuario> servicoUsuario;
         private readonly Mock<IMediator> mediator;
 
@@ -28,7 +27,6 @@ namespace SME.SGP.Aplicacao.Teste.Comandos
             servicoAutenticacao = new Mock<IServicoAutenticacao>();
             servicoUsuario = new Mock<IServicoUsuario>();
             servicoPerfil = new Mock<IServicoPerfil>();
-            servicoTokenJwt = new Mock<IServicoTokenJwt>();
             servicoEOL = new Mock<IServicoEol>();
             var repositorioCache = new Mock<IRepositorioCache>();
             servicoAbrangencia = new Mock<IServicoAbrangencia>();
@@ -38,7 +36,7 @@ namespace SME.SGP.Aplicacao.Teste.Comandos
             repositorioSuporteUsuario = new Mock<IRepositorioSuporteUsuario>();
             mediator = new Mock<IMediator>();
 
-            comandosUsuario = new ComandosUsuario(servicoAutenticacao.Object, servicoUsuario.Object, servicoPerfil.Object, servicoEOL.Object, servicoTokenJwt.Object,
+            comandosUsuario = new ComandosUsuario(servicoAutenticacao.Object, servicoUsuario.Object, servicoPerfil.Object, servicoEOL.Object,
                 repositorioCache.Object, servicoAbrangencia.Object, repositorioHistoricoEmailUsuario.Object, 
                 repositorioSuporteUsuario.Object, mediator.Object);
         }
